+++ src/OpenTelemetry/Logs/LoggerProviderSdk.cs
@@ -194,7 +194,12 @@ internal sealed class LoggerProviderSdk : LoggerProvider
     }
 
     /// <inheritdoc />
-    protected override bool TryCreateLogger(
+#if EXPOSE_EXPERIMENTAL_FEATURES
+    protected
+#else
+    internal
+#endif
+        override bool TryCreateLogger(
         string? name,
 #if NETSTANDARD2_1_OR_GREATER || NET6_0_OR_GREATER
         [NotNullWhen(true)]