+++ src/OpenTelemetry.Extensions.Hosting/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* The experimental APIs previously covered by `OTEL1000`
+  (`OpenTelemetryBuilder.WithLogging` method) will now be part of the public API
+  and supported in stable builds.
+  ([#5648](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5648))
+
 ## 1.9.0-alpha.1
 
 Released 2024-May-20